pyproject.toml
[project]
name = "receipt-index"
version = "0.1.0"
description = "Pre-index receipts from email, extract metadata, and provide CLI search for accounting reconciliation."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "python-slugify>=8.0",
    "weasyprint>=62.0",
]

[project.scripts]
receipt-index = "receipt_index.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "detect-secrets>=1.5",
    "pip-audit>=2.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/receipt_index"]

[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
# Pydantic models need these imports at runtime for type resolution
"src/receipt_index/models.py" = ["TC003"]
# pydantic-ai Agent uses Pydantic output types that need runtime imports
"src/receipt_index/extraction.py" = ["TC003"]

[tool.ruff.lint.isort]
known-first-party = ["receipt_index"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["slugify.*", "weasyprint.*", "pydantic_ai.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=receipt_index --cov-report=term-missing --cov-report=html"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["receipt_index"]
omit = [
    # Stub modules — not yet implemented
    "src/receipt_index/cli.py",
    "src/receipt_index/db.py",
]

[tool.coverage.report]
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "pass",
    "@overload",
    "\\.\\.\\.",
]
